Pagination - full refactor
