+++ src/pages/register/index.tsx
@@ -1,4 +1,4 @@
-import { Box, Stack, VStack } from "@chakra-ui/react"
+import { Box, Button, Stack, VStack } from "@chakra-ui/react"
 import { useWeb3React } from "@web3-react/core"
 import NotConnectedError from "components/admin/common/NotConnectedError"
 import useSubmitCommunityData from "components/admin/hooks/useSubmitCommunityData"
@@ -46,12 +46,17 @@ const Page = (): JSX.Element => {
       <Box sx={generatedColors}>
         <Layout title="Integrate token">
           <Stack spacing={{ base: 7, xl: 9 }}>
-            <Pagination
-              isAdminPage
-              isCommunityTabDisabled
-              onSaveClick={methods.handleSubmit(onSubmit)}
-              saveBtnLoading={loading}
-            />
+            <Pagination isCommunityTabDisabled>
+              <Button
+                isLoading={loading}
+                variant="solid"
+                colorScheme="primary"
+                size="md"
+                onClick={methods.handleSubmit(onSubmit)}
+              >
+                Save
+              </Button>
+            </Pagination>
             <VStack spacing={12}>
               <Details />
               <Appearance
